ex4: PR review fix
