+++ src/ex4/.gitignore
@@ -1,1 +1,3 @@
 node_modules
+tasks
+task.txt
